requirements.txt
streamlit
python-docx 
PyPDF2
zipfile
tempfile
pyinstaller